+++ events/src/events/multi/TemperatureEventHandler.java
@@ -1,5 +1,5 @@
 package events.multi;
 
 public interface TemperatureEventHandler {
-	public void onNewTemperaturValue(float temperature);
+	public void onNewTemperaturValue(int temperature);
 }
