+++ events/src/events/single/Node.java
@@ -9,9 +9,9 @@ import events.single.EventHandler;
 public class Node {
 	private List<Timeout> timeoutHandler = new ArrayList<Timeout>();
 	private List<EventHandler> temperatureHandler = new ArrayList<EventHandler>();
-	private float temperature;
+	private int temperature;
 	
-	public Node(float temperature) {
+	public Node(int temperature) {
 		this.temperature = temperature;
 	}
 	
@@ -27,7 +27,7 @@ public class Node {
 		}
 	}
 	
-	public void receiveNewTemperature(float temperature) {
+	public void receiveNewTemperature(int temperature) {
 		if (this.temperature != temperature) {
 			System.out.println("Node: temperature value change from " + this.temperature + " to: " + temperature);
 			this.temperature = temperature;
