+++ events/src/events/single/TemperatureEvent.java
@@ -1,8 +1,8 @@
 package events.single;
 
 public class TemperatureEvent extends Event {
-	public float temperture;
-	public TemperatureEvent(float temperature) {
+	public int temperture;
+	public TemperatureEvent(int temperature) {
 		this.temperture = temperature;
 	}
 }
